moved unused modules to sample directory
