--- src/modules/droppoint-shipping/store/index.ts
@@ -1,7 +1,0 @@
-import { Module } from 'vuex'
-import { actions } from './actions'
-
-export const module: Module<any, any> = {
-  namespaced: true,
-  actions
-}
