+++ src/modules/payment-cash-on-delivery/index.ts
@@ -2,33 +2,44 @@ import { StorefrontModule } from '@vue-storefront/module';
 import { isServer } from '@vue-storefront/core/helpers'
 import Vue from 'vue';
 import InfoComponent from './components/Info.vue'
+import EventBus from '@vue-storefront/core/compatibility/plugins/event-bus'
 
 export const PaymentCashOnDeliveryModule: StorefrontModule = function (app, store, router, moduleConfig, appConfig) {
   // Place the order. Payload is empty as we don't have any specific info to add for this payment method '{}'
   let correctPaymentMethod = false
   const placeOrder = () => {
     if (correctPaymentMethod) {
-      Vue.prototype.$bus.$emit('checkout-do-placeOrder', {})
+      EventBus.$emit('checkout-do-placeOrder', {})
     }
   }
-
+  // Update the methods
+  let paymentMethodConfig = {
+    'title': 'Cash on delivery',
+    'code': 'cashondelivery',
+    'cost': 0,
+    'costInclTax': 0,
+    'default': true,
+    'offline': true,
+    'is_server_method': false
+  }
+  store.dispatch('payment/addMethod', paymentMethodConfig)
   if (!isServer) {
     // Update the methods
     let paymentMethodConfig = {
       'title': 'Cash on delivery',
       'code': 'cashondelivery',
       'cost': 0,
-      'costInclTax': 0,
+      'cost_incl_tax': 0,
       'default': true,
       'offline': true,
       'is_server_method': false
     }
     store.dispatch('payment/addMethod', paymentMethodConfig)
 
-    Vue.prototype.$bus.$on('checkout-before-placeOrder', placeOrder)
+    EventBus.$on('checkout-before-placeOrder', placeOrder)
 
     // Mount the info component when required.
-    Vue.prototype.$bus.$on('checkout-payment-method-changed', (paymentMethodCode) => {
+    EventBus.$on('checkout-payment-method-changed', (paymentMethodCode) => {
       let methods = store.state['payment-backend-methods'].methods
       if (methods) {
         let method = methods.find(item => (item.code === paymentMethodCode))
