--- src/modules/sample-custom-entity-graphql/testentity.gql
@@ -1,8 +1,0 @@
-query testentity ($filter: TestInput) {
-  testentity(
-    filter: $filter
-  )
-  {
-    hits
-  }
-}
